pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "newspaper3k>=0.2.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "streamlit>=1.56.0",
]
